+++ modules/engine/src/index.ts
@@ -159,8 +159,12 @@ export class VectorEngine implements IVectorEngine {
     if (!valid) {
       return Result.fail(new Error(validate.errors?.map((err) => err.message).join(",")));
     }
-    const channel = await this.vector.getChannelState(params.channelAddress);
-    return Result.ok(channel);
+    try {
+      const channel = await this.store.getChannelState(params.channelAddress);
+      return Result.ok(channel);
+    } catch (e) {
+      return Result.fail(e);
+    }
   }
 
   private async getTransferState(
@@ -205,8 +209,13 @@ export class VectorEngine implements IVectorEngine {
     if (!valid) {
       return Result.fail(new Error(validate.errors?.map((err) => err.message).join(",")));
     }
-    const transfer = await this.store.getTransferByRoutingId(params.channelAddress, params.routingId);
-    return Result.ok(transfer);
+
+    try {
+      const transfer = await this.store.getTransferByRoutingId(params.channelAddress, params.routingId);
+      return Result.ok(transfer);
+    } catch (e) {
+      return Result.fail(e);
+    }
   }
 
   private async getTransferStatesByRoutingId(
@@ -217,8 +226,12 @@ export class VectorEngine implements IVectorEngine {
     if (!valid) {
       return Result.fail(new Error(validate.errors?.map((err) => err.message).join(",")));
     }
-    const transfers = await this.store.getTransfersByRoutingId(params.routingId);
-    return Result.ok(transfers);
+    try {
+      const transfers = await this.store.getTransfersByRoutingId(params.routingId);
+      return Result.ok(transfers);
+    } catch (e) {
+      return Result.fail(e);
+    }
   }
 
   private async getChannelStateByParticipants(
@@ -234,8 +247,12 @@ export class VectorEngine implements IVectorEngine {
     if (!valid) {
       return Result.fail(new Error(validate.errors?.map((err) => err.message).join(",")));
     }
-    const channel = await this.vector.getChannelStateByParticipants(params.alice, params.bob, params.chainId);
-    return Result.ok(channel);
+    try {
+      const channel = await this.vector.getChannelStateByParticipants(params.alice, params.bob, params.chainId);
+      return Result.ok(channel);
+    } catch (e) {
+      return Result.fail(e);
+    }
   }
 
   private async getChannelStates(): Promise<
@@ -244,8 +261,12 @@ export class VectorEngine implements IVectorEngine {
       Error | OutboundChannelUpdateError
     >
   > {
-    const channel = await this.vector.getChannelStates();
-    return Result.ok(channel);
+    try {
+      const channel = await this.store.getChannelStates();
+      return Result.ok(channel);
+    } catch (e) {
+      return Result.fail(e);
+    }
   }
 
   private async getRegisteredTransfers(
@@ -282,7 +303,7 @@ export class VectorEngine implements IVectorEngine {
 
     const chainProviders = this.chainService.getChainProviders();
     if (chainProviders.isError) {
-      return Result.fail(new Error(chainProviders.getError()!.message));
+      return Result.fail(chainProviders.getError()!);
     }
 
     const setupRes = await this.vector.setup({
@@ -379,7 +400,11 @@ export class VectorEngine implements IVectorEngine {
       return Result.fail(new Error(validate.errors?.map((err) => err.message).join(",")));
     }
 
-    const channel = await this.store.getChannelState(params.channelAddress);
+    const channelRes = await this.getChannelState({ channelAddress: params.channelAddress });
+    if (channelRes.isError) {
+      return Result.fail(channelRes.getError()!);
+    }
+    const channel = channelRes.getValue();
     if (!channel) {
       return Result.fail(
         new OutboundChannelUpdateError(OutboundChannelUpdateError.reasons.ChannelNotFound, params as any),
@@ -408,7 +433,11 @@ export class VectorEngine implements IVectorEngine {
       return Result.fail(new Error(validate.errors?.map((err) => err.message).join(",")));
     }
 
-    const channel = await this.store.getChannelState(params.channelAddress);
+    const channelRes = await this.getChannelState({ channelAddress: params.channelAddress });
+    if (channelRes.isError) {
+      return Result.fail(channelRes.getError()!);
+    }
+    const channel = channelRes.getValue();
     if (!channel) {
       return Result.fail(
         new OutboundChannelUpdateError(OutboundChannelUpdateError.reasons.ChannelNotFound, params as any),
@@ -444,14 +473,16 @@ export class VectorEngine implements IVectorEngine {
       return Result.fail(new Error(validate.errors?.map((err) => err.message).join(",")));
     }
 
-    // TODO: consider a store method to find active transfer by routingId
-    const transfer = await this.store.getTransferState(params.transferId);
+    const transferRes = await this.getTransferState({ transferId: params.transferId });
+    if (transferRes.isError) {
+      return Result.fail(transferRes.getError()!);
+    }
+    const transfer = transferRes.getValue();
     if (!transfer) {
       return Result.fail(
         new OutboundChannelUpdateError(OutboundChannelUpdateError.reasons.TransferNotFound, params as any),
       );
     }
-    // TODO validate that transfer hasn't already been resolved?
 
     // First, get translated `create` params using the passed in conditional transfer ones
     const resolveResult = convertResolveConditionParams(params, transfer);
@@ -476,7 +507,11 @@ export class VectorEngine implements IVectorEngine {
       return Result.fail(new Error(validate.errors?.map((err) => err.message).join(",")));
     }
 
-    const channel = await this.store.getChannelState(params.channelAddress);
+    const channelRes = await this.getChannelState({ channelAddress: params.channelAddress });
+    if (channelRes.isError) {
+      return Result.fail(channelRes.getError()!);
+    }
+    const channel = channelRes.getValue();
     if (!channel) {
       return Result.fail(
         new OutboundChannelUpdateError(OutboundChannelUpdateError.reasons.ChannelNotFound, params as any),